refact(form-fields): reusing form fields and restructuring services feature
