+++ dashboard/src/features/services/components/dialogs/delete-confirmation.dialog.tsx
