+++ dashboard/src/features/services/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./form-fields";
+export * from "./tables";
+export * from "./dialogs";
