+++ dashboard/src/features/services/components/tables/index.ts
@@ -1,0 +1,4 @@
+export * from './users'
+export * from './inbounds'
+export * from './services'
+export * from './services/columns'
