+++ dashboard/src/features/services/services/index.ts
@@ -1,5 +1,7 @@
-export * from './services.query'
-export * from './service.query'
-export * from './create-service.mutate'
-export * from './delete-service.mutate'
-export * from './update-service.mutate'
+export * from "./create-service.mutate";
+export * from "./delete-service.mutate";
+export * from "./service-inbounds.query";
+export * from "./service-users.query";
+export * from "./service.query";
+export * from "./services.query";
+export * from "./update-service.mutate";
