--- dashboard/src/features/services/tables/index.ts
@@ -1,4 +1,0 @@
-// export * from './users.table'
-// export * from './inbounds.table'
-export * from './services'
-export * from './services/columns'
