+++ dashboard/src/features/users/components/dialogs/mutation/fields/index.ts
@@ -1,8 +1,6 @@
 export * from "./data-limit"
 export * from "./expire-date"
 export * from "./note"
-export * from "./services"
 export * from "./data-limit-reset-strategy"
-export * from "./username"
 export * from "./activation-deadline"
 export * from "./usage-duration"
