+++ dashboard/src/features/users/components/index.ts
@@ -1,6 +1,6 @@
 export * from './users-status'
 export * from './services-card'
-export * from './date-field'
+export * from './form-fields'
 export * from './online-status'
 export * from './used-user-traffic'
 export * from "./tables"
